Исправил ошибки по модуляции
